Allow the popover-position breakpoint to be specified with a variable
This commit adds a "$popover-bottom-breakpoint" variable. If set to a
width, this value is the window width below which popovers will be shown
at the bottom of the screen. If set to "false", popovers will never be
shown at the bottom, no matter how narrow the screen is. The default
value is 720 pixels, which matches the previous Littlefoot behavior--
this change is completely backward compatible.