Fix bottom padding issue on Firefox
Due to Firefox bug 748518 [1]--which is unfortunately marked INVALID--
Firefox calculates the bottom padding of the footnote content
differently from other browsers. The upshot is that, in any Littlefoot
footnotes that are long enough to scroll, there is not enough padding
below the content, and the last line of the content is partially
obscured by the "fade out" gradient at the bottom of the footnote. This
commit fixes the issue for me (in Firefox 67 on macOS) and maintains the
previous (good) appearance under Safari 12 and Chrome 75.

Fixes #21.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=748518