Fix a crash when there is an empty footnote
Previously, when backlinks were removed from the popover copy of the
footnote, it would stop removing ancestors based on when the element
was empty. However, with an empty footnote, it would never stop and
would throw an exception when the parent element was null, breaking all
footnotes on the page.